+++ src/Assets/Scenes/Main.unity
@@ -194,15 +194,16 @@ GameObject:
   - component: {fileID: 243286406}
   - component: {fileID: 243286408}
   - component: {fileID: 243286404}
+  - component: {fileID: 243286422}
   - component: {fileID: 243286409}
   - component: {fileID: 243286421}
   - component: {fileID: 243286420}
   - component: {fileID: 243286419}
   - component: {fileID: 243286418}
   - component: {fileID: 243286417}
+  - component: {fileID: 243286414}
   - component: {fileID: 243286416}
   - component: {fileID: 243286415}
-  - component: {fileID: 243286414}
   - component: {fileID: 243286413}
   - component: {fileID: 243286412}
   - component: {fileID: 243286411}
@@ -278,7 +279,7 @@ Transform:
   m_LocalScale: {x: 0.75, y: 0.75, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &243286404
 MonoBehaviour:
@@ -292,26 +293,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e090fe64f4d4f624f89fff53e60c2320, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  colliders:
-  - {fileID: 243286421}
-  - {fileID: 243286420}
-  - {fileID: 243286419}
-  - {fileID: 243286418}
-  - {fileID: 243286417}
-  - {fileID: 243286416}
-  - {fileID: 243286415}
-  - {fileID: 243286414}
-  - {fileID: 243286413}
-  - {fileID: 243286412}
-  - {fileID: 243286411}
-  - {fileID: 243286410}
-  - {fileID: 243286407}
-  currentColliderIndex: 0
   horizontalMoveSpeed: 5
   rotationSpeed: 10
   maxRotation: 45
   xBounds: 2.5
-  animator: {fileID: 0}
 --- !u!50 &243286405
 Rigidbody2D:
   serializedVersion: 4
@@ -1145,10 +1130,11 @@ PolygonCollider2D:
       - {x: -0.45549315, y: -0.2673528}
       - {x: -0.7392482, y: -0.26995993}
       - {x: -0.80632424, y: -0.2791383}
-      - {x: -0.8487937, y: -0.324968}
-      - {x: -0.8634584, y: -0.35419145}
-      - {x: -0.8690787, y: -0.39105105}
-      - {x: -0.8677134, y: -0.43065196}
+      - {x: -0.8373367, y: -0.32600942}
+      - {x: -0.84679365, y: -0.3531497}
+      - {x: -0.85154885, y: -0.38433635}
+      - {x: -0.8503309, y: -0.39417562}
+      - {x: -0.8479241, y: -0.4285688}
       - {x: -0.8524615, y: -0.48421383}
       - {x: -0.8406548, y: -0.5439622}
       - {x: -0.82880557, y: -0.6106111}
@@ -1168,7 +1154,7 @@ PolygonCollider2D:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 243286401}
-  m_Enabled: 1
+  m_Enabled: 0
   m_Density: 1
   m_Material: {fileID: 0}
   m_IsTrigger: 0
@@ -1186,13 +1172,13 @@ PolygonCollider2D:
   m_AutoTiling: 0
   m_Points:
     m_Paths:
-    - - {x: 0.90625, y: -0.64375}
-      - {x: 0.9375, y: -0.59375}
-      - {x: 0.94375, y: -0.54375}
-      - {x: 0.94375, y: -0.44375}
-      - {x: 0.90000004, y: -0.35625002}
-      - {x: 0.83750004, y: -0.29375002}
-      - {x: 0.76875, y: -0.24375}
+    - - {x: 0.8153536, y: -0.61591476}
+      - {x: 0.8303816, y: -0.5551065}
+      - {x: 0.8324094, y: -0.5100786}
+      - {x: 0.82615227, y: -0.3785794}
+      - {x: 0.8122338, y: -0.33698925}
+      - {x: 0.7901059, y: -0.30048105}
+      - {x: 0.7401424, y: -0.24711552}
       - {x: 0.675, y: -0.21875}
       - {x: 0.5625, y: -0.20625}
       - {x: 0.425, y: -0.19375001}
@@ -1220,16 +1206,16 @@ PolygonCollider2D:
       - {x: -0.275, y: -0.0125}
       - {x: -0.25, y: -0.11875}
       - {x: -0.3625, y: -0.175}
-      - {x: -0.6375, y: -0.2125}
-      - {x: -0.73125, y: -0.22500001}
-      - {x: -0.81875, y: -0.28125}
-      - {x: -0.89375, y: -0.3625}
-      - {x: -0.93125004, y: -0.43125}
-      - {x: -0.9375, y: -0.49375}
-      - {x: -0.9375, y: -0.55625004}
-      - {x: -0.90000004, y: -0.6375}
-      - {x: -0.85, y: -0.73125}
-      - {x: -0.76875, y: -0.8}
+      - {x: -0.6130634, y: -0.23388201}
+      - {x: -0.6625221, y: -0.24638201}
+      - {x: -0.72711277, y: -0.28125033}
+      - {x: -0.7624034, y: -0.33195445}
+      - {x: -0.7937943, y: -0.40223157}
+      - {x: -0.82142633, y: -0.46625906}
+      - {x: -0.8214262, y: -0.4951586}
+      - {x: -0.8190539, y: -0.58862674}
+      - {x: -0.7980722, y: -0.6686314}
+      - {x: -0.7198768, y: -0.778618}
       - {x: -0.675, y: -0.825}
       - {x: -0.5375, y: -0.83125}
       - {x: -0.36875, y: -0.8625}
@@ -1258,9 +1244,9 @@ PolygonCollider2D:
       - {x: 0.3125, y: -0.86875004}
       - {x: 0.35, y: -0.86875004}
       - {x: 0.45000002, y: -0.84375}
-      - {x: 0.6, y: -0.83750004}
-      - {x: 0.71875, y: -0.8}
-      - {x: 0.78125, y: -0.76875}
+      - {x: 0.58912235, y: -0.8157442}
+      - {x: 0.6678844, y: -0.7825891}
+      - {x: 0.7328637, y: -0.7272226}
 --- !u!60 &243286416
 PolygonCollider2D:
   m_ObjectHideFlags: 0
@@ -1823,6 +1809,32 @@ PolygonCollider2D:
       - {x: -0.00625, y: -0.3125}
       - {x: 0, y: -0.33125}
       - {x: 0.00625, y: -0.32500002}
+--- !u!114 &243286422
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 243286401}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8d920d76887bd2646bccc7eda69413f9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  colliders:
+  - {fileID: 243286421}
+  - {fileID: 243286420}
+  - {fileID: 243286419}
+  - {fileID: 243286418}
+  - {fileID: 243286417}
+  - {fileID: 243286416}
+  - {fileID: 243286415}
+  - {fileID: 243286415}
+  - {fileID: 243286413}
+  - {fileID: 243286412}
+  - {fileID: 243286411}
+  - {fileID: 243286410}
+  - {fileID: 243286407}
 --- !u!224 &275277322
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2479,7 +2491,7 @@ Camera:
   far clip plane: 1000
   field of view: 60
   orthographic: 1
-  orthographic size: 5.3357143
+  orthographic size: 5.3333335
   m_Depth: -1
   m_CullingMask:
     serializedVersion: 2
@@ -2507,7 +2519,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &615335748
 GameObject:
@@ -2666,11 +2678,11 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 624523392}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 7.335294, z: 0}
+  m_LocalPosition: {x: 0, y: 7.3333335, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &624523395
 MonoBehaviour:
@@ -2830,8 +2842,9 @@ RectTransform:
   - {fileID: 334854163}
   - {fileID: 1399010772}
   - {fileID: 769970020}
+  - {fileID: 615335749}
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -2911,7 +2924,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1166127882
 MonoBehaviour:
@@ -3362,7 +3375,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6824043117476076021, guid: d44a5c901f220dd41a22541718e3ed76, type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 6824043117476076021, guid: d44a5c901f220dd41a22541718e3ed76, type: 3}
       propertyPath: m_AnchorMax.x
@@ -3760,7 +3773,7 @@ Transform:
   m_LocalScale: {x: 0.5, y: 0.5, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1932853541
 GameObject:
@@ -4067,5 +4080,5 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
