+++ src/Assets/Scripts/Obstacles/HurtBox.cs
@@ -9,16 +9,14 @@ public class HurtBox : MonoBehaviour
     [SerializeField]
     private GameObject effect;
 
-    private float deathTime = 80;
-
     private void OnTriggerEnter2D(Collider2D other){
         if (other.CompareTag("Player")) {           
             if (effect != null){
                 Instantiate(effect, transform.position, Quaternion.identity);               
             }            
-                ScreenShakeController.instance.TriggerShake(0.5f);
-                GameController.instance.DecrementHealth(damage);
-                Destroy(gameObject, deathTime);
+            ScreenShakeController.instance.TriggerShake(0.5f);
+            GameController.instance.DecrementHealth(damage);
+            Destroy(gameObject);
         }
     }   
 }
